+++ packages/client-ts/src/context/profiles/profile.ts
@@ -34,7 +34,7 @@ export class Profile extends EventEmitter implements IProfile {
    *
    * **Do not instantiate directly.**
    *
-   * Access the current user's profile via {@link App.profile}
+   * Access the current user's profile
    *
    * @constructor
    */