+++ packages/vda-xp-reward-client/src/blockchain/userApi.ts
@@ -1,8 +1,21 @@
+<<<<<<< HEAD
 import { VdaClientConfig, Web3SelfTransactionConfig } from '@verida/types'
 import { ethers, BytesLike, BigNumber } from "ethers";
 import { getContractInfoForBlockchainAnchor, getVeridaSign } from "@verida/vda-common";
 import { VeridaClientBase } from "@verida/vda-client-base";
 import { VeridaTokenClient } from '@verida/vda-token-client';
+=======
+import {
+    getVeridaContract,
+    VeridaContract
+} from "@verida/web3"
+import { Web3SelfTransactionConfig, VdaClientConfig } from '@verida/types'
+import { ethers, Contract, BytesLike } from "ethers";
+import { getContractInfoForNetwork, RPC_URLS, getVeridaSign, getVeridaSignWithNonce } from "@verida/vda-common";
+import { JsonRpcProvider } from '@ethersproject/providers';
+import { explodeDID } from '@verida/helpers'
+import EncryptionUtils from "@verida/encryption-utils";
+>>>>>>> 3b0946c30b18bb8815a5a89aeac04b595b2d6beb
 
 /**
  * This is the claim information
@@ -27,10 +40,62 @@ export interface ClaimXPInfo {
 /**
  * This is a client class that interacts with the `VDAXPReward` contract of Verida
  */
+<<<<<<< HEAD
 export class VeridaXPRewardClient extends VeridaClientBase{
 
     public constructor(config: VdaClientConfig) {
         super(config, "xpReward");
+=======
+export class VeridaXPRewardClient {
+
+    protected config: VdaClientConfig
+    protected network: string
+    protected didAddress?: string
+
+    protected vdaWeb3Client? : VeridaContract
+
+    protected readOnly: boolean
+    protected contract?: ethers.Contract
+
+    public constructor(config: VdaClientConfig) {
+        if (!config.callType) {
+            config.callType = 'web3'
+        }
+
+        this.config = config
+        this.readOnly = true
+        if (!config.web3Options) {
+            config.web3Options = {}
+        }
+
+        this.network = config.network
+
+        if (config.callType == 'web3' && !(<Web3SelfTransactionConfig>config.web3Options).rpcUrl) {
+            (<Web3SelfTransactionConfig> config.web3Options).rpcUrl = <string> RPC_URLS[this.network]
+        }
+
+        const contractInfo = getContractInfoForNetwork("VDAXPReward", this.network)
+
+        if (config.did) {
+            this.readOnly = false
+            const { address } = explodeDID(config.did)
+            this.didAddress = address.toLowerCase()
+            
+            this.vdaWeb3Client = getVeridaContract(
+                config.callType, 
+                {...contractInfo,
+                ...config.web3Options})
+        } else {
+            let rpcUrl = (<Web3SelfTransactionConfig>config.web3Options).rpcUrl
+            if (!rpcUrl) {
+                rpcUrl = <string> RPC_URLS[this.network]
+            }
+
+            const provider = new JsonRpcProvider(rpcUrl)
+
+            this.contract = new Contract(contractInfo.address, contractInfo.abi.abi, provider)
+        }
+>>>>>>> 3b0946c30b18bb8815a5a89aeac04b595b2d6beb
     }
 
     /**
@@ -158,12 +223,17 @@ export class VeridaXPRewardClient extends VeridaClientBase{
             claims,
             requestSignature,
             requestProof
+<<<<<<< HEAD
         );
+=======
+        )
+>>>>>>> 3b0946c30b18bb8815a5a89aeac04b595b2d6beb
 
         if (response.success !== true) {
             throw new Error(`Failed to claim xp: (${response.reason})`)
         }
     }
+<<<<<<< HEAD
 
     /**
      * Deposit verida token to the `XPRewardContract` for `claimXPReward()` function calls
@@ -200,4 +270,6 @@ export class VeridaXPRewardClient extends VeridaClientBase{
             console.log("Errr");
         }
     }
+=======
+>>>>>>> 3b0946c30b18bb8815a5a89aeac04b595b2d6beb
 }