+++ packages/vda-xp-reward-client/test/user_read.test.ts
@@ -1,8 +1,16 @@
 require('dotenv').config();
+<<<<<<< HEAD
 import { getBlockchainAPIConfiguration } from "@verida/vda-common-test"
 import { addInitialData } from "./helpers";
 import { VeridaXPRewardClient } from "../src/blockchain/userApi";
 import { Test_BlockchainAnchor } from "./const";
+=======
+import { DID_LIST, getBlockchainAPIConfiguration, ERC20Manager } from "@verida/vda-common-test"
+import { EnvironmentType } from "@verida/types";
+import { addInitialData } from "./helpers";
+import { Wallet } from 'ethers';
+import { VeridaXPRewardClient } from "../src/blockchain/userApi";
+>>>>>>> 3b0946c30b18bb8815a5a89aeac04b595b2d6beb
 
 const assert = require('assert')